# docs/swagger.yaml
basePath: /
definitions:
  frameworkdto.CreatedResponseDTO:
    description: Created response structure
    properties:
      data: {}
      message:
        example: Resource created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  frameworkdto.ErrorDetailDTO:
    description: Error detail structure
    properties:
      code:
        example: ERR_001
        type: string
      details: {}
      message:
        example: An error occurred
        type: string
    type: object
  frameworkdto.ErrorResponseDTO:
    description: Error API response structure
    properties:
      error:
        $ref: '#/definitions/frameworkdto.ErrorDetailDTO'
      success:
        example: false
        type: boolean
    type: object
  frameworkdto.GetLicenceTypeDTO:
    properties:
      description:
        type: string
      id:
        type: integer
      max_seats:
        type: integer
      name:
        type: string
    type: object
...
  frameworkdto.LicenceTypeCreateRequestDTO:
    properties:
      description:
        type: string
      max_seats:
        type: integer
      name:
        type: string
    type: object
  frameworkdto.LicenceTypeUpdateRequestDTO:
    properties:
      description:
        type: string
      id:
        type: integer
      max_seats:
        type: integer
      name:
        type: string
    type: object
  frameworkdto.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  frameworkdto.LoginResponseDTO:
    properties:
      token:
        type: string
    type: object
...
  frameworkdto.SuccessResponseDTO:
    description: Successful API response structure
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@geekible.com
    name: API Support
  description: |-
    A comprehensive service framework API with multi-tenant support, user management, authentication, and authorization.

    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`

    ## Error Responses
    All error responses follow a standard format with success, error code, and message fields.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Service Framework API
  version: "1.0"
paths:
...
  /authentication/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/frameworkdto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/frameworkdto.LoginResponseDTO'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      summary: User login
      tags:
      - Authentication
  /licence-type/create:
    post:
      consumes:
      - application/json
      description: Create a new licence type (requires authentication, super admin
        only)
      parameters:
      - description: Licence type creation details
        in: body
        name: createLicenceTypeDTO
        required: true
        schema:
          $ref: '#/definitions/frameworkdto.LicenceTypeCreateRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Licence type created successfully
          schema:
            $ref: '#/definitions/frameworkdto.CreatedResponseDTO'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Create a new licence type
      tags:
      - Licence Type
  /licence-type/delete:
    delete:
      consumes:
      - application/json
      description: Delete a licence type by ID (requires authentication, super admin
        only)
      parameters:
      - description: Licence Type ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Licence type deleted successfully
          schema:
            $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
        "400":
          description: ID is required or Invalid ID
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Delete a licence type
      tags:
      - Licence Type
  /licence-type/get-all:
    get:
      consumes:
      - application/json
      description: Get list of all licence types (requires authentication, super admin
        only)
      produces:
      - application/json
      responses:
        "200":
          description: Licence types fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/frameworkdto.GetLicenceTypeDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get all licence types
      tags:
      - Licence Type
  /licence-type/get-by-id:
    get:
      consumes:
      - application/json
      description: Get a specific licence type by ID (requires authentication, super
        admin only)
      parameters:
      - description: Licence Type ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Licence type fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/frameworkdto.GetLicenceTypeDTO'
              type: object
        "400":
          description: ID is required or Invalid ID
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get licence type by ID
      tags:
      - Licence Type
  /licence-type/update:
    put:
      consumes:
      - application/json
      description: Update an existing licence type (requires authentication, super
        admin only)
      parameters:
      - description: Licence type update details
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/frameworkdto.LicenceTypeUpdateRequestDTO'
      produces:
      - application/json
      responses:
        "202":
          description: Licence type updated successfully
          schema:
            $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Update a licence type
      tags:
      - Licence Type
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"